SUFTP Sharability & SUFTP status update
That sounds weird but I changed some things which allow the SUFTP.h file to be compiled when Qt is available and when it's not. It's important for keeping the protocol in sync with changes between the CLI and GUI versions of SyncUp. Also added a slot and signal for the GUI to update its status label based on SUFTP's status.